Fix: Addresses can be 64bits now
Should fix the negative address display with recent firmwares

+++ src/PropertiesWidget.cpp
@@ -131,7 +131,7 @@ void PropertiesWidget::setType(ChimeraTK::DataDescriptor const& dataDescriptor) 
   }
 }
 
-void PropertiesWidget::setAddress(int bar, int address, int sizeInBytes) {
+void PropertiesWidget::setAddress(int bar, uint64_t address, int sizeInBytes) {
   ui.numericalAddressGroupBox->show();
   ui.barDisplay->setText(QString::number(bar));
   ui.addressDisplay->setText(QString::number(address));
