+++ src/PropertiesWidget.h
@@ -25,7 +25,7 @@ class PropertiesWidget : public QWidget {
  protected:
   // Numerical addressed -only methods of the interface
   void setFixedPointInfo(int width, int nFracBits, bool signedFlag, unsigned int nDimensions);
-  void setAddress(int bar, int address, int sizeInBytes);
+  void setAddress(int bar, uint64_t address, int sizeInBytes);
   // other internal methods to avoid screens of unstructured code...
   void setShape(unsigned int nDimensions, unsigned int nChannels, unsigned int nElements);
   void setType(ChimeraTK::DataDescriptor const& dataDescriptor);
